RoomSearchContext 통합
